refactor: 코드리뷰 반영하여 수정
