--- src/constants/CHANNEL.js
@@ -1,6 +1,0 @@
-module.exports = {
-  JOIN: "join",
-  LEAVE: "leave",
-  USER_SEND: "user-send",
-  BROADCAST: "broadcast",
-};
