--- src/constants/MESSAGE.js
@@ -1,5 +1,0 @@
-module.exports = {
-  UNAUTHORIZED: "Unauthorized",
-  INVALID_INPUT: "Invalid input",
-  LIMITED_PROJECT: "The number of projects has reached the limit",
-};
