+++ src/constants/index.js
@@ -1,0 +1,14 @@
+module.exports = {
+  MESSAGE: {
+    UNAUTHORIZED: "Unauthorized",
+    INVALID_INPUT: "Invalid input",
+    LIMITED_PROJECT: "The number of projects has reached the limit",
+  },
+  CHANNEL: {
+    JOIN: "join",
+    LEAVE: "leave",
+    USER_SEND: "user-send",
+    BROADCAST: "broadcast",
+  },
+  LIMITED_PROJECT_COUNT: 3,
+};
