+++ src/routes/project.route.js
@@ -5,6 +5,6 @@ const router = express.Router();
 const verifyToken = require("./middleware/verifyToken");
 const projectController = require("./controllers/project.controller");
 
-router.route("/").post(verifyToken, projectController.saveProject);
+router.route("/").post(verifyToken, projectController.createProject);
 
 module.exports = router;
